+++ pleo-antaeus-data/src/main/kotlin/io/pleo/antaeus/data/AntaeusDal.kt
@@ -23,8 +23,8 @@ class AntaeusDal(db: Database) {
     fun updateInvoiceStatus(id: Int, newStatus: InvoiceStatus): Invoice? = invoiceDal.updateInvoiceStatus(id, newStatus)
     fun fetchInvoicePagesByStatus(status: InvoiceStatus, pageSize: Int?, marker: Int?): InvoicePage =
         invoiceDal.fetchInvoicesPageByStatus(status, pageSize.let { DEFAULT_PAGE_SIZE }, marker)
-    fun fetchInvoicePagesByCustomer(customer: Int, pageSize: Int?, marker: Int?): InvoicePage =
-        invoiceDal.fetchInvoicesPageByCustomer(customer, pageSize.let { DEFAULT_PAGE_SIZE }, marker)
+    fun fetchInvoicePagesByCustomer(customer: Int, status: InvoiceStatus, pageSize: Int?, marker: Int?): InvoicePage =
+        invoiceDal.fetchInvoicesPageByCustomer(customer, status, pageSize.let { DEFAULT_PAGE_SIZE }, marker)
 
 
     fun fetchCustomer(id: Int): Customer? = customerDal.fetchCustomer(id)